Move to_chars_float implementation details into src folder
